Refactor server recipe to use Pebble server
This is a major refactor for the server recipe which is used for testing only.
This migrates us away from the clunky and slow Boulder server to the lightweight
and nimble Pebble server [1]. This largely uses the method that the upstream
acme cookbook [2] uses for it's own testing however with a few modifications.

Namely, instead of compiling the pebble binary during an initial converge, pull
in a pre-compiled binary that exists on our own servers. This means we do not
need to pull in any additional build dependencies to get this going. While
Pebble does not require the use of a local DNS server for validation purposes, I
decided to keep this in there to help with integration testing with InSpec.

This pulls in the same test cookbook and tests that the
osl-letsencrypt-boulder-server cookbook had for its setup with a few minor
modifications.

Some changes to keep in mind:
- Attributes for ``node['boulder']['host_aliases']`` should be renamed to
  ``node['osl-acme']['pebble']['host_aliases']``
- Setting ``node['acme']['dir']`` in ``.kitchen.yml`` is no longer required as
  it's set in the server recipe by default.
- Support for was re-added CentOS 6 which required creating a proper init
  script.

[1] https://github.com/letsencrypt/pebble
[2] https://github.com/schubergphilis/chef-acme/blob/master/test/fixtures/cookbooks/acme_server/recipes/default.rb